+++ docs/grafana/rmf-app/oxygen-webhelp/app/search/index/keywords.js
@@ -1,4 +1,4 @@
-define(function() {var keywords=[{w:"Accessing",p:["p0","p13"]},{w:"the",p:["p0","p5","p6","p10","p18"]},{w:"Grafana",p:["p0","p5","p7","p8","p10","p13","p18","p20"]},{w:"dashboard",p:["p0","p4"]},{w:"Introduction",p:["p1"]},{w:"to",p:["p1","p2"]},{w:"Alerts",p:["p1"]},{w:"Applying",p:["p2"]},{w:"visualization",p:["p2","p20"]},{w:"RMF",p:["p2","p3","p4","p10","p15","p18","p19","p20"]},{w:"data",p:["p2","p3","p9"]},{w:"Creating",p:["p3"]},{w:"sources",p:["p3"]},{w:"master",p:["p4"]},{w:"Defining",p:["p5"]},{w:"server",p:["p5"]},{w:"Error",p:["p6"]},{w:"reporting",p:["p6"]},{w:"in",p:["p6","p20"]},{w:"plugin",p:["p6","p10","p18"]},{w:"configuration",p:["p7"]},{w:"parameters",p:["p7"]},{w:"through",p:["p8"]},{w:"IBM",p:["p8","p11","p15"]},{w:"z/OS",p:["p8","p10","p18"]},{w:"Management",p:["p8"]},{w:"Facility",p:["p8"]},{w:"Historical",p:["p9"]},{w:"collection",p:["p9"]},{w:"Installing",p:["p10"]},{w:"for",p:["p10","p13","p18"]},{w:"on",p:["p10","p13","p18"]},{w:"Integrating",p:["p11"]},{w:"DDS",p:["p11"]},{w:"Dashboards",p:["p11","p13","p14"]},{w:"with",p:["p11"]},{w:"Z",p:["p11"]},{w:"OMEGAMON",p:["p11"]},{w:"PDF",p:["p12"]},{w:"Guide",p:["p12"]},{w:"Prerequisites",p:["p13"]},{w:"z/OSMF",p:["p13"]},{w:"Prometheus",p:["p14"]},{w:"Sample",p:["p14"]},{w:"query",p:["p15","p19"]},{w:"languages",p:["p15"]},{w:"Release",p:["p16"]},{w:"notes",p:["p16"]},{w:"Troubleshooting",p:["p17"]},{w:"issues",p:["p17"]},{w:"Upgrading",p:["p18"]},{w:"Variable",p:["p19"]},{w:"syntax",p:["p19"]},{w:"of",p:["p20"]},{w:"Monitor",p:["p20"]},{w:"III",p:["p20"]},{w:"metrics",p:["p20"]}];
+define(function() {var keywords=[{w:"Accessing",p:["p0","p14"]},{w:"the",p:["p0","p5","p6","p11","p19"]},{w:"Grafana",p:["p0","p5","p7","p8","p11","p14","p19","p21"]},{w:"dashboard",p:["p0","p4"]},{w:"Introduction",p:["p1"]},{w:"to",p:["p1","p2"]},{w:"Alerts",p:["p1"]},{w:"Applying",p:["p2"]},{w:"visualization",p:["p2","p21"]},{w:"RMF",p:["p2","p3","p4","p10","p11","p16","p19","p20","p21"]},{w:"data",p:["p2","p3","p9"]},{w:"Creating",p:["p3"]},{w:"sources",p:["p3"]},{w:"master",p:["p4"]},{w:"Defining",p:["p5"]},{w:"server",p:["p5"]},{w:"Error",p:["p6"]},{w:"reporting",p:["p6"]},{w:"in",p:["p6","p10","p21"]},{w:"plugin",p:["p6","p11","p19"]},{w:"configuration",p:["p7"]},{w:"parameters",p:["p7"]},{w:"through",p:["p8"]},{w:"IBM",p:["p8","p12","p16"]},{w:"z/OS",p:["p8","p11","p19"]},{w:"Management",p:["p8"]},{w:"Facility",p:["p8"]},{w:"Historical",p:["p9"]},{w:"collection",p:["p9"]},{w:"Importing",p:["p10"]},{w:"Dashboards",p:["p10","p12","p14","p15"]},{w:"created",p:["p10"]},{w:"Performance",p:["p10"]},{w:"Monitoring",p:["p10"]},{w:"Installing",p:["p11"]},{w:"for",p:["p11","p14","p19"]},{w:"on",p:["p11","p14","p19"]},{w:"Integrating",p:["p12"]},{w:"DDS",p:["p12"]},{w:"with",p:["p12"]},{w:"Z",p:["p12"]},{w:"OMEGAMON",p:["p12"]},{w:"PDF",p:["p13"]},{w:"Guide",p:["p13"]},{w:"Prerequisites",p:["p14"]},{w:"z/OSMF",p:["p14"]},{w:"Prometheus",p:["p15"]},{w:"Sample",p:["p15"]},{w:"query",p:["p16","p20"]},{w:"languages",p:["p16"]},{w:"Release",p:["p17"]},{w:"notes",p:["p17"]},{w:"Troubleshooting",p:["p18"]},{w:"issues",p:["p18"]},{w:"Upgrading",p:["p19"]},{w:"Variable",p:["p20"]},{w:"syntax",p:["p20"]},{w:"of",p:["p21"]},{w:"Monitor",p:["p21"]},{w:"III",p:["p21"]},{w:"metrics",p:["p21"]}];
 var ph={};
 ph["p0"]=[0, 1, 2, 3];
 ph["p1"]=[4, 5, 6];
@@ -10,17 +10,18 @@ ph["p6"]=[16, 17, 18, 1, 19];
 ph["p7"]=[2, 20, 21];
 ph["p8"]=[2, 22, 23, 24, 25, 26];
 ph["p9"]=[27, 10, 28];
-ph["p10"]=[29, 1, 9, 30, 24, 19, 31, 2];
-ph["p20"]=[8, 53, 9, 54, 55, 56, 18, 2];
-ph["p12"]=[38, 39];
-ph["p11"]=[32, 33, 34, 35, 23, 36, 37, 34];
-ph["p14"]=[42, 43, 34];
-ph["p13"]=[40, 30, 0, 2, 34, 31, 41];
-ph["p16"]=[46, 47];
-ph["p15"]=[23, 9, 44, 45];
-ph["p18"]=[50, 1, 9, 30, 24, 19, 31, 2];
-ph["p17"]=[48, 49];
-ph["p19"]=[9, 51, 44, 52];
+ph["p10"]=[29, 30, 31, 18, 9, 32, 33];
+ph["p21"]=[8, 57, 9, 58, 59, 60, 18, 2];
+ph["p20"]=[9, 55, 48, 56];
+ph["p12"]=[37, 38, 30, 39, 23, 40, 41, 30];
+ph["p11"]=[34, 1, 9, 35, 24, 19, 36, 2];
+ph["p14"]=[44, 35, 0, 2, 30, 36, 45];
+ph["p13"]=[42, 43];
+ph["p16"]=[23, 9, 48, 49];
+ph["p15"]=[46, 47, 30];
+ph["p18"]=[52, 53];
+ph["p17"]=[50, 51];
+ph["p19"]=[54, 1, 9, 35, 24, 19, 36, 2];
      return {
          keywords: keywords,
          ph: ph